build(env): create a conda env for use with Emme
Manage conda environments with separate scripts that can be called to
activate an environment when needed. Environments are automatically
created if they don't exist. Replace use of a path to where Python is
installed with Emme.

The Emme Python libraries are included in a conda environment that can
be extended - unlike the one installed with Emme. The Emme Python
libraries require the same Python version as Emme, which is currently
different than the Python version used by cmap_modedest. Therefore, two
conda environments are needed at this time.